+++ app/Harvest/SpiceHarvesterService.php
@@ -36,29 +36,23 @@ class SpiceHarvesterService
      * @param \DateTime $to
      */
     public function harvest(SpiceHarvesterHarvest $harvest, \DateTime $from = null, \DateTime $to = null, $only_ids = []) {
-        $timeStart = time();
-        $result = new Result();
-
-        $this->harvesters[$harvest->type]->harvest($harvest, $result, $from, $to, $only_ids);
-
-        $harvest->status = SpiceHarvesterHarvest::STATUS_COMPLETED;
-        $harvest->completed = date('Y-m-d H:i:s');
-        $harvest->status_messages = trans('harvest.status_messages.completed', [
-            'processed' => $result->getTotal(),
-            'created' => $result->getInserted(),
-            'updated' => $result->getUpdated(),
-            'deleted' => $result->getDeleted(),
-            'skipped' => $result->getSkipped(),
-            'time' => time() - $timeStart,
-        ]);
-        $harvest->status_messages .= PHP_EOL . implode(PHP_EOL, $result->getErrorMessages());
-        $harvest->save();
+        $this->harvesters[$harvest->type]->tryHarvest($harvest, $from, $to, $only_ids);
     }
 
     /**
      * @param SpiceHarvesterRecord $record
      */
     public function harvestSingle(SpiceHarvesterRecord $record) {
-        $this->harvesters[$record->type]->harvestSingle($record, new Result());
+        $this->harvesters[$record->type]->tryHarvestSingle($record, new Result());
+    }
+
+    /**
+     * @param SpiceHarvesterHarvest $harvest
+     */
+    public function harvestFailed(SpiceHarvesterHarvest $harvest) {
+        $failed = $harvest->records()->failed()->get();
+        foreach ($failed as $record) {
+            $this->harvestSingle($record);
+        }
     }
 }