+++ CHANGELOG.md
@@ -3,8 +3,10 @@ All notable changes to this project will be documented in this file[^1].
 
 ## [Unreleased]
 ### Added
+- color picker in the catalog
 - override docker php container for tests
 - ImageCarousel Component
+- harvesting of failed records
 
 ### Fixed
 - correct authority zero-value years to null
@@ -210,6 +212,8 @@ All notable changes to this project will be documented in this file[^1].
 ### Added
 - Color-list component on artwork detail
 - Color description
+- LESS to CSS compilation
+- ColorPicker component
 - Zoom-viewer component
 - Multiple images per artwork
 - Moravian gallery importer